move score calculation into actor_utils
