+++ x/emissions/module/rewards/rewards_test.go
@@ -2207,7 +2207,7 @@ func (s *RewardsTestSuite) TestOnlyFewTopActorsGetReward() {
 	networkLossBundles, err := s.emissionsKeeper.GetNetworkLossBundleAtBlock(s.ctx, topicId, block)
 	s.Require().NoError(err)
 
-	infererScores, err := rewards.GenerateInferenceScores(
+	infererScores, err := actorutils.GenerateInferenceScores(
 		s.ctx,
 		s.emissionsKeeper,
 		topicId,
@@ -2215,7 +2215,7 @@ func (s *RewardsTestSuite) TestOnlyFewTopActorsGetReward() {
 		*networkLossBundles)
 	s.Require().NoError(err)
 
-	forecasterScores, err := rewards.GenerateForecastScores(
+	forecasterScores, err := actorutils.GenerateForecastScores(
 		s.ctx,
 		s.emissionsKeeper,
 		topicId,
